pyproject.toml
[project]
name = "rag-doc-bot-examensarbete"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "lancedb>=0.26.0",
    "pandas>=2.3.3",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
